% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotviewR.R
\name{pivotviewR}
\alias{pivotviewR}
\title{pivotviewR}
\usage{
pivotviewR(df, CollectionName, ImgBase = "", colFormats = NULL,
  filterVisible = NULL, metadataVisible = NULL, searchable = NULL,
  width = NULL, height = NULL)
}
\arguments{
\item{df}{the dataframe, should have these columns [Name, Description, Img, Href, Id]}

\item{CollectionName}{what your collection is called}

\item{ImgBase}{where your image files are located}

\item{colFormats}{format strings for the data, [n, p, m] = [number, percentage, money] followed by the number of decimal places.. i.e. "n2" means number with two decimal places}

\item{filterVisible}{list of true/false (for each column in df)}

\item{metadataVisible}{list of true/false (for each column in df)}

\item{searchable}{list of true/false (for each column in df)}

\item{width}{the width}

\item{height}{the height}
}
\description{
Create a new pivotviewer object from a data.frame, which should include the following columns:
\itemize{
\item Name: The name of the item
\item Description: a longer text description the displays when zoomed in on an item
\item Img: the filename of image, should be relative to ImgBase
\item Href: a webpage you want to link the item to
\item Id: optional unique item identifier
}
}

